Separate Mutations into different component, clean up Get
